cmd/utils/rls/gen_hasura_rls_test.go
package rls_test

import (
	"bytes"
	"fmt"
	"io/ioutil"
	"testing"

	"github.com/manabie-com/backend/cmd/utils/rls"
	mock_fileio "github.com/manabie-com/backend/mock/golibs/io"

	"github.com/stretchr/testify/assert"
	"gopkg.in/yaml.v2"
)

func getMockData(fileInput string, fileExpected string) ([]byte, []byte, error) {
	errChan := make(chan error, 2)
	fileInputChan := make(chan []byte)
	fileExpectedChan := make(chan []byte)
	go func() {
		content, err := ioutil.ReadFile(fileInput)
		fileInputChan <- content
		errChan <- err
	}()
	go func() {
		content, err := ioutil.ReadFile(fileExpected)
		fileExpectedChan <- content
		errChan <- err
	}()

	return <-fileInputChan, <-fileExpectedChan, <-errChan
}

func parseFileYaml(file string) ([]rls.HasuraTable, error) {
	tables := []rls.HasuraTable{}

	content, err := ioutil.ReadFile(file)
	if err != nil {
		return nil, err
	}

	err = yaml.Unmarshal(content, &tables)

	if err != nil {
		return nil, err
	}
	return tables, nil
}

func getMockDataHasuraV2(fileInput string, fileExpected string) ([]rls.HasuraTable, []rls.HasuraTable, error) {
	errChan := make(chan error, 2)
	fileInputChan := make(chan []rls.HasuraTable)
	fileExpectedChan := make(chan []rls.HasuraTable)
	go func() {
		content, err := parseFileYaml(fileInput)
		fileInputChan <- content
		errChan <- err
	}()
	go func() {
		content, err := parseFileYaml(fileExpected)
		fileExpectedChan <- content
		errChan <- err
	}()

	return <-fileInputChan, <-fileExpectedChan, <-errChan
}

type RLSTestCase struct {
	name           string
	command        interface{}
	expectedErrStr string
	args           []string
	setup          func() error
}

type MockLessonInput struct {
	mockFileIO              *mock_fileio.MockFileUtils
	inputFile               string
	outputFile              string
	expectedHasuraV2Folder  string
	tableFileHasuraV2File   string
	expectedStrGrantedTable string
	expectedGrantedFile     string
	fileMockGrantedView     string
}

func setupLessonMockFunc(input MockLessonInput) error {
	lessonContentInput, lessonContentExpected, err := getMockDataHasuraV2(input.inputFile, input.outputFile)
	lessonFileInput, err := yaml.Marshal(lessonContentInput[0])
	lessonFileExpected, err := yaml.Marshal(lessonContentExpected[0])

	fileName := "public_lessons.yaml"
	fileNames := []string{fileName}
	expectedHasuraV2File := fmt.Sprintf("%s/%s", input.expectedHasuraV2Folder, fileName)
	fileGrantedPermission := fmt.Sprintf("../../../mock/testing/testdata/rls/%s.yaml", input.fileMockGrantedView)
	expectedGrantedFileContent, err := ioutil.ReadFile(fileGrantedPermission)

	input.mockFileIO.On("AppendStrToFile", input.tableFileHasuraV2File, input.expectedStrGrantedTable).Once().Return(nil, nil)

	input.mockFileIO.On("GetFileNamesOnDir", input.expectedHasuraV2Folder).Once().Return(fileNames, nil)
	input.mockFileIO.On("GetFileContent", expectedHasuraV2File).Once().Return(lessonFileInput, nil)

	input.mockFileIO.On("WriteFile", input.expectedGrantedFile, expectedGrantedFileContent).Once().Return(nil, nil)
	input.mockFileIO.On("WriteFile", expectedHasuraV2File, lessonFileExpected).Once().Return(nil, nil)

	return err
}

func TestHasuraRootCmd(t *testing.T) {
	mockFileIO := &mock_fileio.MockFileUtils{}
	rlsHasura := &rls.Hasura{
		IOUtils: mockFileIO,
	}
	messageError := "actual is not expected"
	rlsTypeFlag := "--rlsType"
	tableFlag := "--table"
	pkeyFlag := "--pkey"
	accessPathTableFlag := "--accessPathTable"
	databaseFlag := "--databaseName"
	permissionPrefixFlag := "--permissionPrefix"
	accessPathTableKeyFlag := "--accessPathTableKey"
	writeHasuraPermissionFlag := "--writePermissionHasura"
	otherTemplateFilterFlag := "--otherTemplateFilterName"
	expectedHasuraFile := "deployments/helm/manabie-all-in-one/charts/bob/files/hasura/metadata/tables.yaml"
	expectedHasuraV2Folder := "deployments/helm/manabie-all-in-one/charts/bob/files/hasurav2/metadata/databases/bob/tables"
	databaseName := "bob"
	hasuraVersionFlag := "--hasuraVersion"
	hasuraVersionV2 := "2"

	classCaseInput := "../../../mock/testing/testdata/rls/class_case.yaml"
	classCaseExpected := "../../../mock/testing/testdata/rls/class_case_expected.yaml"
	classCaseAllRolesExpected := "../../../mock/testing/testdata/rls/class_case_all_roles_expected.yaml"
	classCaseAllRolesWithOwnersExpected := "../../../mock/testing/testdata/rls/class_case_all_roles_owners_expected.yaml"
	classCaseWithOwnerExpected := "../../../mock/testing/testdata/rls/class_case_with_owner_expected.yaml"
	studentCaseInput := "../../../mock/testing/testdata/rls/student_case.yaml"
	studentCaseExpected := "../../../mock/testing/testdata/rls/student_case_expected.yaml"
	studentCaseWithOwnerExpected := "../../../mock/testing/testdata/rls/student_case_with_owner_expected.yaml"
	studentCaseAllRolesExpected := "../../../mock/testing/testdata/rls/student_case_all_roles_expected.yaml"
	studentCaseAllRolesTemplate3Expected := "../../../mock/testing/testdata/rls/student_case_all_roles_template_3_expected.yaml"
	lessonCaseInput := "../../../mock/testing/testdata/rls/lesson_case.yaml"
	lessonCaseExpected := "../../../mock/testing/testdata/rls/lesson_case_expected.yaml"
	lessonCaseWithOwnerExpected := "../../../mock/testing/testdata/rls/lesson_case_with_owner_expected.yaml"
	expectedGrantedFile := fmt.Sprintf("%s/%s", expectedHasuraV2Folder, "public_granted_permissions.yaml")
	userPrefixPermission := "user.student"
	tableFileHasuraV2File := fmt.Sprintf("%s/%s", expectedHasuraV2Folder, "tables.yaml")
	expectedStrGrantedTable := `- "!include public_granted_permissions.yaml"`

	templateVersionFlag := "--templateVersion"
	ownerFlag := "--ownerCol"

	userClassPrefixPermission := "user.class"
	userLessonPrefixPermission := "user.lesson"

	addRLSToAllPermissionHasuraFlag := "--addRLSToAllPermissionHasura"

	testCases := []RLSTestCase{
		{
			name:           "Should added relationship and RLS of Hasura when location inside input table",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "class", pkeyFlag, "location_id", databaseFlag, databaseName, permissionPrefixFlag, userClassPrefixPermission},
			setup: func() error {
				classContentInput, classContentExpected, err := getMockData(classCaseInput, classCaseExpected)

				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(classContentInput, nil)
				mockFileIO.On("WriteFile", expectedHasuraFile, classContentExpected).Once().Return(nil, nil)

				return err
			},
		},
		{
			name:           "Should added relationship and RLS of Hasura when location outside input table",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "students", pkeyFlag, "student_id", databaseFlag, databaseName, permissionPrefixFlag, userPrefixPermission, accessPathTableFlag, "user_access_paths", accessPathTableKeyFlag, "location_id"},
			setup: func() error {
				studentContentInput, studentContentExpected, err := getMockData(studentCaseInput, studentCaseExpected)

				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(studentContentInput, nil)
				mockFileIO.On("WriteFile", expectedHasuraFile, studentContentExpected).Once().Return(nil, nil)
				return err
			},
		},
		{
			name:           "Should added relationship and RLS of Hasura when location inside input table and location column named center_id",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "lessons", pkeyFlag, "lesson_id", databaseFlag, databaseName, permissionPrefixFlag, userLessonPrefixPermission, accessPathTableKeyFlag, "center_id"},
			setup: func() error {
				lessonContentInput, lessonContentExpected, err := getMockData(lessonCaseInput, lessonCaseExpected)

				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(lessonContentInput, nil)
				mockFileIO.On("WriteFile", expectedHasuraFile, lessonContentExpected).Once().Return(nil, nil)
				return err
			},
		},
		{
			name:           "Should replace if existed RLS",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "students", pkeyFlag, "student_id", databaseFlag, databaseName, permissionPrefixFlag, userPrefixPermission, accessPathTableFlag, "user_access_paths", accessPathTableKeyFlag, "location_id"},
			setup: func() error {
				input, expected, err := getMockData(studentCaseExpected, studentCaseExpected)

				mockFileIO.On("WriteFile", expectedHasuraFile, expected).Once().Return(nil, nil)
				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(input, nil)

				return err
			},
		},
		{
			name:           "Should added relationship and RLS of Hasura version 2 when location inside input table",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "class", pkeyFlag, "location_id", databaseFlag, databaseName, permissionPrefixFlag, userClassPrefixPermission, hasuraVersionFlag, hasuraVersionV2},
			setup: func() error {
				classContentInput, classContentExpected, err := getMockDataHasuraV2(classCaseInput, classCaseExpected)
				classInput, err := yaml.Marshal(classContentInput[0])
				byteExpected, err := yaml.Marshal(classContentExpected[0])

				fileName := "public_class.yaml"
				fileNames := []string{fileName}
				expectedHasuraV2File := fmt.Sprintf("%s/%s", expectedHasuraV2Folder, fileName)
				fileGrantedPermission := "../../../mock/testing/testdata/rls/class_granted_permissions.yaml"
				expectedGrantedFileContent, err := ioutil.ReadFile(fileGrantedPermission)

				mockFileIO.On("GetFileNamesOnDir", expectedHasuraV2Folder).Once().Return(fileNames, nil)
				mockFileIO.On("AppendStrToFile", tableFileHasuraV2File, expectedStrGrantedTable).Once().Return(nil, nil)
				mockFileIO.On("GetFileContent", expectedHasuraV2File).Once().Return(classInput, nil)
				mockFileIO.On("WriteFile", expectedGrantedFile, expectedGrantedFileContent).Once().Return(nil, nil)
				mockFileIO.On("WriteFile", expectedHasuraV2File, byteExpected).Once().Return(nil, nil)
				return err
			},
		},
		{
			name:           "Should added relationship and RLS of Hasura version 2 when location outside input table",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "students", pkeyFlag, "student_id", databaseFlag, databaseName, permissionPrefixFlag, userPrefixPermission, accessPathTableFlag, "user_access_paths", accessPathTableKeyFlag, "location_id", hasuraVersionFlag, hasuraVersionV2},
			setup: func() error {
				studentContentInput, studentContentExpected, err := getMockDataHasuraV2(studentCaseInput, studentCaseExpected)
				studentFileInput, err := yaml.Marshal(studentContentInput[0])
				userFileInput, err := yaml.Marshal(studentContentInput[1])
				accessPathInput, err := yaml.Marshal(studentContentInput[2])
				studentFileContentExpected, err := yaml.Marshal(studentContentExpected[0])
				accessPathFileContentExpected, err := yaml.Marshal(studentContentExpected[2])

				studentTableFileName := "public_students.yaml"
				userTableFileName := "public_users.yaml"
				accessPathFileName := "public_user_access_paths.yaml"

				fileNames := []string{studentTableFileName, userTableFileName, accessPathFileName}
				studentFile := fmt.Sprintf("%s/%s", expectedHasuraV2Folder, studentTableFileName)
				userFile := fmt.Sprintf("%s/%s", expectedHasuraV2Folder, userTableFileName)
				accessPathFile := fmt.Sprintf("%s/%s", expectedHasuraV2Folder, accessPathFileName)
				fileGrantedPermission := "../../../mock/testing/testdata/rls/students_granted_permissions.yaml"
				expectedGrantedFileContent, err := ioutil.ReadFile(fileGrantedPermission)

				mockFileIO.On("AppendStrToFile", tableFileHasuraV2File, expectedStrGrantedTable).Once().Return(nil, nil)

				mockFileIO.On("GetFileNamesOnDir", expectedHasuraV2Folder).Once().Return(fileNames, nil)

				mockFileIO.On("GetFileContent", studentFile).Once().Return(studentFileInput, nil)
				mockFileIO.On("GetFileContent", userFile).Once().Return(userFileInput, nil)
				mockFileIO.On("GetFileContent", accessPathFile).Once().Return(accessPathInput, nil)

				mockFileIO.On("WriteFile", expectedGrantedFile, expectedGrantedFileContent).Once().Return(nil, nil)
				mockFileIO.On("WriteFile", accessPathFile, accessPathFileContentExpected).Once().Return(nil, nil)
				mockFileIO.On("WriteFile", studentFile, studentFileContentExpected).Once().Return(nil, nil)

				return err
			},
		},
		{
			name:           "Should added relationship and RLS of Hasura version 2 when location inside input table and location column named center_id",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "lesson", pkeyFlag, "lesson_id", databaseFlag, databaseName, permissionPrefixFlag, userLessonPrefixPermission, accessPathTableKeyFlag, "center_id", hasuraVersionFlag, hasuraVersionV2},
			setup: func() error {
				mockInput := MockLessonInput{
					mockFileIO:              mockFileIO,
					inputFile:               lessonCaseInput,
					outputFile:              lessonCaseExpected,
					expectedHasuraV2Folder:  expectedHasuraV2Folder,
					tableFileHasuraV2File:   tableFileHasuraV2File,
					expectedStrGrantedTable: expectedStrGrantedTable,
					expectedGrantedFile:     expectedGrantedFile,
					fileMockGrantedView:     "lesson_granted_permissions",
				}
				return setupLessonMockFunc(mockInput)
			},
		},
		{
			name:           "Should added correct filter for select permission template 4",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "class", pkeyFlag, "location_id", databaseFlag, databaseName, permissionPrefixFlag, userClassPrefixPermission, templateVersionFlag, "4", ownerFlag, "owners"},
			setup: func() error {
				classContentInput, classContentExpected, err := getMockData(classCaseInput, classCaseWithOwnerExpected)

				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(classContentInput, nil)
				mockFileIO.On("WriteFile", expectedHasuraFile, classContentExpected).Once().Return(nil, nil)

				return err
			},
		},
		{
			name:           "Should add template 4 beside with template 1 have location inside table",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "students", databaseFlag, databaseName, templateVersionFlag, "4", ownerFlag, "owners", otherTemplateFilterFlag, "user"},
			setup: func() error {
				studentContentInput, studentContentExpected, err := getMockData(studentCaseExpected, studentCaseWithOwnerExpected)

				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(studentContentInput, nil)
				mockFileIO.On("WriteFile", expectedHasuraFile, studentContentExpected).Once().Return(nil, nil)

				return err
			},
		},
		{
			name:           "Should add template 4 beside with template 1 have location outside table",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "lessons", databaseFlag, databaseName, templateVersionFlag, "4", ownerFlag, "owners", otherTemplateFilterFlag, "lessons_location_permission"},
			setup: func() error {
				lessonContentInput, lessonContentExpected, err := getMockData(lessonCaseExpected, lessonCaseWithOwnerExpected)

				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(lessonContentInput, nil)
				mockFileIO.On("WriteFile", expectedHasuraFile, lessonContentExpected).Once().Return(nil, nil)
				return err
			},
		},
		{
			name:           "Should added relationship and RLS of Hasura version 2 when location inside input table and location column named center_id with template 4",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "lessons", databaseFlag, databaseName, hasuraVersionFlag, hasuraVersionV2, templateVersionFlag, "4", ownerFlag, "owners", otherTemplateFilterFlag, "lessons_location_permission"},
			setup: func() error {
				mockInput := MockLessonInput{
					mockFileIO:              mockFileIO,
					inputFile:               lessonCaseExpected,
					outputFile:              lessonCaseWithOwnerExpected,
					expectedHasuraV2Folder:  expectedHasuraV2Folder,
					tableFileHasuraV2File:   tableFileHasuraV2File,
					expectedStrGrantedTable: expectedStrGrantedTable,
					expectedGrantedFile:     expectedGrantedFile,
					fileMockGrantedView:     "lesson_granted_permissions_owner_expected",
				}
				return setupLessonMockFunc(mockInput)
			},
		},
		{
			name:           "Should throw error when missing owner col in template version 4",
			expectedErrStr: "Error: ownerCol is required in template Version 4\n",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "lessons", pkeyFlag, "lesson_id", databaseFlag, databaseName, permissionPrefixFlag, userLessonPrefixPermission, accessPathTableKeyFlag, "center_id", templateVersionFlag, "4"},
			setup: func() error {
				return nil
			},
		},
		{
			name:           "Should added relationship and RLS of Hasura when location column in access path table name center_id",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "lesson_members", pkeyFlag, "lesson_id", databaseFlag, databaseName, permissionPrefixFlag, "user.lesson", accessPathTableFlag, "lessons", accessPathTableKeyFlag, "center_id"},
			setup: func() error {
				inputFile := "../../../mock/testing/testdata/rls/lesson_members_case.yaml"
				expectedFile := "../../../mock/testing/testdata/rls/lesson_members_expected.yaml"
				lessonMemberContentInput, lessonMemberContentExpected, err := getMockData(inputFile, expectedFile)
				mockFileIO.On("WriteFile", expectedHasuraFile, lessonMemberContentExpected).Once().Return(nil, nil)
				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(lessonMemberContentInput, nil)
				return err
			},
		},
		{
			name:           "Should create INSERT/SELECT/UPDATE/DELETE policy when location_id column inside main table",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", tableFlag, "class", pkeyFlag, "location_id", databaseFlag, databaseName, permissionPrefixFlag, userClassPrefixPermission, writeHasuraPermissionFlag, "INSERT/UPDATE/DELETE"},
			setup: func() error {
				classMutationCaseExpected := "../../../mock/testing/testdata/rls/class_case_mutation_expected.yaml"
				classContentInput, classContentExpected, err := getMockData(classCaseInput, classMutationCaseExpected)

				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(classContentInput, nil)
				mockFileIO.On("WriteFile", expectedHasuraFile, classContentExpected).Once().Return(nil, nil)

				return err
			},
		},
		{
			name: "Should create INSERT/SELECT/UPDATE/DELETE policy when location_id column outside main table",
			args: []string{rlsTypeFlag, "hasura", tableFlag, "students", pkeyFlag, "student_id", databaseFlag, databaseName, permissionPrefixFlag, userPrefixPermission, accessPathTableFlag, "user_access_paths", accessPathTableKeyFlag, "location_id", writeHasuraPermissionFlag, "INSERT/UPDATE/DELETE"},
			setup: func() error {
				studentCaseMutationExpected := "../../../mock/testing/testdata/rls/student_case_mutation_expected.yaml"
				studentContentInput, studentContentExpected, err := getMockData(studentCaseInput, studentCaseMutationExpected)
				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(studentContentInput, nil)
				mockFileIO.On("WriteFile", expectedHasuraFile, studentContentExpected).Once().Return(nil, nil)

				return err
			},
		},
		{
			name:           "Should added relationship and RLS of Hasura to all roles when location inside input table",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", addRLSToAllPermissionHasuraFlag, "true", tableFlag, "class", pkeyFlag, "location_id", databaseFlag, databaseName, permissionPrefixFlag, userClassPrefixPermission},
			setup: func() error {
				classContentInput, classContentExpected, err := getMockData(classCaseInput, classCaseAllRolesExpected)

				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(classContentInput, nil)
				mockFileIO.On("WriteFile", expectedHasuraFile, classContentExpected).Once().Return(nil, nil)

				return err
			},
		},
		{
			name:           "Should added relationship and RLS of Hasura to all roles when location inside input table with template 4",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", addRLSToAllPermissionHasuraFlag, "true", tableFlag, "class", pkeyFlag, "location_id", databaseFlag, databaseName, permissionPrefixFlag, userClassPrefixPermission, templateVersionFlag, "4", ownerFlag, "owners"},
			setup: func() error {
				classContentInput, classContentExpected, err := getMockData(classCaseInput, classCaseAllRolesWithOwnersExpected)
				mockFileIO.On("WriteFile", expectedHasuraFile, classContentExpected).Once().Return(nil, nil)
				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(classContentInput, nil)

				return err
			},
		},
		{
			name:           "Should added relationship and RLS of Hasura to all roles when location outside input table",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", addRLSToAllPermissionHasuraFlag, tableFlag, "students", pkeyFlag, "student_id", databaseFlag, databaseName, permissionPrefixFlag, userPrefixPermission, accessPathTableFlag, "user_access_paths", accessPathTableKeyFlag, "location_id"},
			setup: func() error {
				studentContentInput, studentContentExpected, err := getMockData(studentCaseInput, studentCaseAllRolesExpected)

				mockFileIO.On("WriteFile", expectedHasuraFile, studentContentExpected).Once().Return(nil, nil)
				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(studentContentInput, nil)

				return err
			},
		},
		{
			name:           "Should added relationship and RLS of Hasura to all roles when template is 3",
			expectedErrStr: "",
			args:           []string{rlsTypeFlag, "hasura", addRLSToAllPermissionHasuraFlag, tableFlag, "students", templateVersionFlag, "3", databaseFlag, databaseName, permissionPrefixFlag, userPrefixPermission},
			setup: func() error {
				studentContentInput, studentContentExpected, err := getMockData(studentCaseInput, studentCaseAllRolesTemplate3Expected)

				mockFileIO.On("WriteFile", expectedHasuraFile, studentContentExpected).Once().Return(nil, nil)
				mockFileIO.On("GetFileContent", expectedHasuraFile).Once().Return(studentContentInput, nil)

				return err
			},
		},
	}

	for _, testCase := range testCases {
		t.Run(testCase.name, func(t *testing.T) {
			// case
			err := testCase.setup()

			error := new(bytes.Buffer)

			cmd := rls.GetCmd(nil, rlsHasura, nil, nil)
			// cmd.SetOut(error)
			cmd.SetErr(error)
			cmd.SetArgs(testCase.args)

			// when
			cmd.Execute()

			// then
			assert.Equal(t, err, nil, messageError)
			assert.Equal(t, testCase.expectedErrStr, error.String(), messageError)
		})
	}
}
